% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_function.R
\name{keyword_function}
\alias{keyword_function}
\title{A Self-made Median Function}
\usage{
keyword_function(xlsxfile, keyword, outname)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to calculate the median from a numeric vector.
}
\examples{
median_function(seq(1:10))
}
\keyword{median}
